Create LICENSE
Have you thought about adding a LISCENSE (either this one or an alternative)? Your project looks great and I'd love to build off it, and without a lisence, I'd be breaching copyright. I'm guessing you're happy with people building off it as it's a public repo. Apologies if that's not the case.
With love,
Josiah